[next]: ensure route handlers aren't skipped when creating .rsc variants (#11808)
The Next.js builder attempts to route RSC requests to their `.rsc` output if it detects the request contains the `RSC` header. We currently skip creating a `.rsc` variant for route handlers because we assumed that there isn't a reason to make an RSC request to a route handler.

However, this means that something as trivial as having a `<Link />` component pointing to a page that is served by the route handler to throw a 500, since the `Link` component would  automatically attempt a prefetch. 

This removes the handling that skipped creating `.rsc` outputs for route handlers.