[tests] tweak CI requirements (#11832)
When we split out the node 16/18 tests, both had the same workflow step
name of "Summary", which is the only required step for merging. The Lint
workflow also has a step with the same name. So, it seems that if any of
these succeeded, the PR was mergeable.

I updated the requirements to require Lint, Node 16, and Node 18
versions of this step in the repo settings. This PR updates the names to
match.