[build-utils] Remove implicit reliance on `@types/glob` for downstream consumers (#11914)
Refactor the `GlobOptions` interface in build-utils to not extend from
the base type in `@types/glob`. The reason for this is because it is not
listed as a peer dependency and caused an issue when a newer version of
`glob` package was present which includes types that don't agree with
the `@types/glob` version (there is no `IOptions` interface).