[cli] Display custom environments in `vc env ls` (#11912)
`vc env ls` would not display custom environment names, and this adds support for that.

Unfortunately, environment variables do not contain the names of custom environments (they only contain custom environment ids). By fetching custom environments alongside environment vars, we can resolve ids to their name.

Also, this command still needs to work for users not flagged into custom environments. The custom environment endpoints returns a 404 if the user is not flagged in, so we catch that and continue without using custom environments.

### Before

<img width="643" alt="Screenshot 2024-07-31 at 3 46 52 PM" src="https://github.com/user-attachments/assets/371a0ebd-ff6b-47b9-84e1-a75d0b256350">

### After
<img width="665" alt="image" src="https://github.com/user-attachments/assets/b25f7732-d132-4f1d-9175-823217cf3607">