[cli] Print the custom environment name in `vc list` (#11965)
<img width="835" alt="Screenshot 2024-08-15 at 19 51 04" src="https://github.com/user-attachments/assets/3d58d444-9f76-4e1b-b295-7585451e820b">

The "Environment" column now shows the custom environment if the deployment was created with one. It also links to said environment for terminals that support hyperlinks. The name of the project at the top is also a hyperlink to the Project page in supported terminals, similar to `vc target ls`.

**Note:** I suggest reviewing with [whitespace hidden](https://github.com/vercel/vercel/pull/11965/files?w=1).