[cli] Add error for `vc deploy --env` with no value (#12002)
In preparation for Custom Environments, there's a realistic possibility that users unintentionally use `--env` when deploying to specify an environment rather than an environment variable to deploy with. Since this is _not_ an error state (env vars can be set as empty strings), we notify the user that they may have meant to specify an environment.