[Logs] use level to determine build log highlight (#12044)
We have moved the build log highlight to api instead of in `front` and
`cli`; as a result, cli and front can both have consistent highlight. We
have already deployed and tested the changes to front behind a feature
flag and we would like to roll out the cli changes in the next version.