[api] Create tarballs of example templates instead of serving through Serverless Function (#12128)
The serverless function has started failing with an EMFILE due to the
vercel/vercel repo `.zip` now containing too many files to be opened in
the Lambda environment.

So instead, remove that Lambda completely and create static tarball
assets at build-time. This is also a lot faster and cheaper, and does
not rely on the state of `main` (so older deployments will continue to
work, unlike the current case).

**tl;dr** URLs such as
[`/api/examples/download/remix.tar.gz`](https://vercel-git-debug-api-examples-download.vercel.sh/api/examples/download/remix.tar.gz)
(used by `vc init`) are now static instead of dynamic.