[cli] provision marketplace resources via cli (#12127)
## Summary

This PR allows Vercel CLI users, in certain conditions, to provision
resources from the Vercel Marketplace without leaving the terminal.

## Background

In order for users to buy products on the Vercel Marketplace they need
to install the integration of the Provider first. This means for the
Vercel CLI there are two scenarios:

### Scenario 1: The Marketplace integration IS NOT installed

In this scenario the user needs to accept the terms for Vercel
Marketplace and the Provider (e.g. Supabase). If this scenario is
detected we offer the user to open the Vercel dashboard in the team and
project context where they can accept the terms and install the product
via the Web UI. This is how `vc install` works today and has not
changed.

### Scenario 2: The Marketplace integration IS installed and the terms
have been accepted

If the terms have been accepted we can attempt to provision the resource
without leaving the terminal. If we determine the provisioning can't be
done in the terminal, we offer to open Vercel dashboard instead (same as
scenario 1).

The conditions are:

* The Marketplace integration was installed once and the terms have been
accepted
* The selected product is a storage product (this is the only
Marketplace product supported at the moment)
* The provider-defined metadata can be expressed in the terminal UI

## How to test

Build and link Vercel CLI locally. The 3 integrations supported today
are

```
vc i supabase
vc i redis
vc i edgedb
```

---------

Co-authored-by: Sean Massa <EndangeredMassa@gmail.com>