[next] Support Request Chaining Types (#12063)
This introduces Request Chaining which is a stabilization of the `experimentalStreamingLambdaPath` feature (for Partial Prerendering) to include additional fields like `headers` which will be sent alongside the request to the "chained" lambda in order to perform the resume request. Previously only the output path was provided, which served as the `x-matched-path` header value as well as the output path. This was at fault, as it assumed that the `basePath` was `/`. When the `basePath` is configured in Next.js, this breaks the implementation. This shifts the logic of which headers to send down to Next.js so it can just provide the explicit headers to be set rather than the logic having to exist upstream.

This is a supporting PR to #12117 which adds the implementation.