[next]: fix defaultLocale redirect when using domains pattern (#12166)
The redirect rule for `defaultLocale` is incorrectly matching the `/index` route resulting in a intermediary rewrite to `/<locale>/index`, which then falls through to the trailing slash redirect. This causes a 308 to the default locale for the domain even though the user would already be on the correct path.

This splits the existing `$wildcard/$1` handling into two separate routes, one for the index route and another for everything else, so that it is handled properly. 

This cannot be tested as it requires assigning a domain. However, I validated this fixes it in [this deployment.](https://next-js-locale-bug-repro.vercel.app/)