Add base telemetry behavior and one demonstration of its use (#12173)
This PR is a the first in a series of PRs intended to add telemetry to
the Vercel CLI. We intended to track events of
* command invocation
* subcommand invocation
* argument use (but not their their actual values)¹
* option/flag use (but not any actual values)¹

The hope is we can use this data to better understand customer usage of
the CLI to:

* remove unused commands
* introduce new commands that might simplify customer workflows
* introduce new features like tab-completion, "did you mean..."
suggestions on mistyped commands, etc.

## The Structure of Telemetry Tracking
This structure is heavily cribbed from two other Vercel projects that
are already tracking metrics: the `next` and `turbo` CLIs. Specifically
we drew heavily from the [`turbo-telemetry`
package](https://github.com/vercel/turborepo/tree/main/packages/turbo-telemetry/src).
However, the Vercel CLI's code structure differs from these two project
that necessitates slightly different code organization.

This PR proposes the following structure:

### The Telemetry Classes
A base telemetry client (`TelemetryBaseClient`) provides a small public
interface to track global command (e.g. `help`) and global option (e.g.
`--cwd`) tracking, some private methods to for tracking specific _kinds_
of events (e.g. "a flag was used") and acts as parent class to
command/subcommand-specific child classes.

The command/subcommand-specific child classes are intended to be defined
and later called within the invocation of each command and subcommand.
The public interface to these classes strictly limits what is allowed to
be tracked at each level and what data we are allowed track.

The subclasses are expected to implement functions that match the
command/argument/option groups used and follow a naming pattern of:

* `trackCliSubcommand<SubcommandName>`
* `trackCliArgument<ArgumentName>`
* `trackCliOption<OptionName>`
* `trackCliFlag<FlagName>`

These methods call to private methods on the base class that format the
data and add a single event to the event store.

### The Event Store
The Vercel CLI is structured with a Russian doll-style nested calls
where the CLI's `main` function calls a outer `<command>` function. This
function calls to an (optional) inner `<subcommand>` function where the
main work occurs. arguments, options, and flags may be parsed at any
layer.

Because we want to track the entirety of an invocation, we need a
structure that is passed to each function call. We currently use a
`client` object for this case and this PR adds a new property of
`telemetryEventStore: TelemetryEventStore` on the client. This store is
created once at the outermost call site and passed as part of client to
each inner function call.

At the end of a cli invocation, we'll have an opportunity to flush these
events to Vercel's existing metrics tracking systems. This flushing is
currently unimplemented and the store's `save` method only logs if in
debug mode.

### Testing
This overall structure creates a nice testing pattern. Our `MockClient`
instance that is passed to command functions under test now has a
`telemetryEventStore` set at initialization to an instance of
`MockTelemetryEventStore`. This mock store is itself a subclass of
`TelemetryEventStore` with an overridden `save` method that is a no-op
and a `reset` method called during test cleanup to zero out any tracked
events accumulated during a run.

This allows us to unit test what the store looks like after a completed
CLI invocation. Right now, this is just using a combination of vitest's
`toMatchObject` and `objectContaining`, but we could write our own
custom matcher or helpers to make this bit more succinct.

This PR includes the tracking added for calls to `vercel domains
transfer-in --code=acode` to offer an example of all of this structure
in use.

---
¹ With a few minor exceptions of which value in a finite collections
that are Vercel defined was passed. For example, we _would_ track which
integration was supplied to `vercel integration add [integration name]`
but would _not_ track which domain name was passed to `vercel domains
add [domain]`.