ECO-585: Handle non-boolean readonly & mark expressions as unsupported (#12336)
## Summary

This PR fixes a bug in the `vc install <integration>` flow. Now
`ui:disabled` is recognized (and skipped), and `ui:hidden` won't break
when set to `create` or `update`. Additionally, complex expressions are
explicitly unsupported.

## Details

Integrations sometimes have complex UI cases (situational hidden,
read-only, disabled) that are allowed in Vercel but weren't properly
detected and handled in the CLI integration installation wizard. The
`ui:disabled` property was being ignored outright in the wizard, while
the `ui:hidden` property was being too aggressive, considering any value
to mean it should be skipped. The `ui:hidden` property can be set to
`update` to mean the field is _not_ readonly during creation, which the
wizard manages. For Upstash Vector, this meant all the necessary fields
were being skipped by the wizard, since they're readonly in update
flows, but not the creation flow.

Additionally, Upstash Vector uses an expression for one of its fields to
determine if it should be visible & populated or not. Our CLI tool does
not have the same full parser for complex expressions that our UI has,
so for cases where we see expressions we should direct the user to the
web UI to complete the process.

## How to test

Build and run locally (or use `pnpm dev` instead of `vc` when in the
`packages/cli` folder).

```
vc install upstash vector
-> select Upstash Vector as the product
```

---------

Co-authored-by: Jeff See <jeffsee.55@gmail.com>