[tests] unskip windows test for `vc dev` using `waitUntil` (#12333)
Reverts vercel/vercel#12267

This test was skipped in Windows because it was flakey. This PR unskips it and resolves the flakiness by adjusting some timeout values.

---

This turned out to be a specific issue with [Node 22 and Windows](https://github.com/nodejs/node/issues/51766) and timers.

The way our use of `waitUntil` works with the `devServer`'s simulations of serverless functions creates an environment where we can kill the dev server at the same time that the `waitUntil` promise expires. I didn't nail down the exact flow that happens here, but this causes the Node 22 + Windows `EPREM` error to bubble up to the test failure.

Ultimately, the specific flow isn't critical to figure out here. This is a test setup issue, not a production issue.

---

In an attempt to prove this is no longer flakey, here are several consecutive successful runs:
- https://github.com/vercel/vercel/actions/runs/11469305192/job/31916358922?pr=12333
- https://github.com/vercel/vercel/actions/runs/11469447637/job/31916775908?pr=12333
- https://github.com/vercel/vercel/actions/runs/11469447637/job/31917078929?pr=12333
- https://github.com/vercel/vercel/actions/runs/11469447637/job/31917252353?pr=12333