[cli] add `VERCEL_ENV` and `VERCEL` to process to simulate runtime (#12358)
In https://github.com/vercel/vercel/pull/11526 we stopped pulling down
environment variables that are set by the runtime into `.env` files
since this really is an appropriate location for them and blank values
can cause issues for some customers.

However, we were relying on these so that `vercel dev` had a similar set
of variables as the platform at runtime. Losing these was a breaking
change.

This PR sets them within the development sever itself to better simulate
our actual runtime environment. I wasn't quite sure of the use of
`projectSettings?.autoExposeSystemEnvs`
[here](https://github.com/vercel/vercel/blob/9537e8c1eb2428e76c8ccf3a52c2e7e8de42e96a/packages/cli/src/util/dev/server.ts#L709-L711)
and potentially the correct location for these is within that `if`
block?

Closes #2878

---------

Co-authored-by: Nathan Rajlich <n@n8.io>