[cli] extract pull invocation from build (#12372)
This is a very basic extraction of the core `vc env pull` logic out of the command invocation for use in the build commmand invocation. This gives us a place to add telemetry later in only the `vc env pull` invocation logic so that invoking `vc build` does not trigger `vc env pull` telemetry metrics.