[cli] Track `git --confirm` earlier (#12419)
Because both `--yes` and `--confirm` are technically allowed we want to track both. To handle the fact that `--confirm` is deprecated is managed with the following pattern: 

```
    output.warn('`--confirm` is deprecated, please use `--yes` instead');
    parsedArgs.flags['--yes'] = parsedArgs.flags['--confirm'];
```

so we need to ensure we've tracked uses of `--confirm` before we smoosh them together. @EndangeredMassa and I audited for this pattern and we've done this in all places _except_ `vercel git`.

`vercel git` needs access to the value in the outer command before the subcommand is called:

https://github.com/vercel/vercel/blob/779caf0b367d6bc865e2151312abba24feac4c5f/packages/cli/src/commands/git/index.ts#L52-L55

So the tracking needs to move.