Add a hidden `telemetry flush` to use for the telemetry send process (#12566)
The way we were previously sending the telemetry event wasn't compatible
with our ability to invoke the CLI from `ts-node` because finding the
`send-telemetry` module wasn't in the expected path. Instead, the
`sendToSubprocess` function invokes the CLI with the `telemetry flush`
command, providing the events as a stringified JSON argument.