Ensure firewall tests run (#12551)
Since we run these tests in chunks, we need to use the `vitest-` scripts
so we know to include them in the "Find changes" action. From what I can
see we haven't actually been testing this package until now.

Also this package isn't designed to run on Node <20 so updated `engines`
and created a new entry in the `chunk-tests` workflow.


https://github.com/vercel/vercel/actions/runs/11730172571/job/32677564246?pr=12551