[next] provide @next/request-context to Next.js handlers (#12286)
Some time ago, Next.js introduced `@next/request-context` which is meant
as a platform-agnostic way of providing `waitUntil` to Next.js:
https://github.com/vercel/next.js/blob/086f715c27391d0559af1b993668b4effc202eaf/packages/next/src/server/after/builtin-request-context.ts#L3-L14

This PR changes the next.js builder to provide this context to handlers.

`@next/request-context` is very close to our existing
`globalThis[Symbol.for("@vercel/request-context")]` (and mirrors a
subset of its interface) but isn't tied to the vercel platform.

---

x-ref: removing usage of `@vercel/request-context` from Next.js (needs
this change)
https://github.com/vercel/next.js/pull/71383