[cli] Normalize POSIX paths in `filePathMap` for `vc build` (#12666)
Fixes an issue when using `vc build` on Windows where the `filePathMap`
object would be written with Windows-style slashes. This caused issues
when trying to create a prebuilt deployment. Using the expected POSIX
style slashes makes the deployment succeed as expected.

Added unit tests to ensure this behavior does not break in the future.