[python] support `VERCEL_IPC_PATH` along `VERCEL_IPC_FD` (#12800)
To communicate between our runtime and language runtimes, we use a Unix
Domain Socket for ICP. We currently open a Unix Domain Socket via a file
descriptor `VERCEL_IPC_FD`, but not all runtimes that we want to support
(e.g. Bun & Deno) support that.

To easily support any language runtime, we're moving to
`VERCEL_IPC_PATH` to let the runtime directly connect to the given IPC
socket, instead of opening it via a FD.

We need to support both ways right now:
- merge this PR that still supports the old way
- enable the new `VERCEL_IPC_PATH` behavior in our runtime
- delete the old `VERCEL_IPC_FD` in the Python runtime