[cli] consolidate error constructors (#12825)
Prior to this PR, we had an `error.ts`, an `errors.ts`, and an
`errors-ts.ts` which made it easy to confuse modules during imports. The
error constructors that were in `errors.ts` bore a lot of similarity to
the constructors in `errors-ts.ts` so this PR merges them and gets rid
of the `errors.ts` file.