[cli] standardize uses of confirm on confirm.ts (#12834)
Prior to this PR, we had a mixed usage of the `confirm` interactive
input component. Some came through `client.input.confirm` and some came
through the function `confirm` that wraps it. This PR standardizes on
the function `confirm` since that was in use in more places and
restricts the options that a call site can use.

This PR doesn't prevent `client.input.confirm` from being called
directly in the future. It may be worth considering an `input-manager`
similar to the previous `output-manager` refactor.