Add video integration category to the CLI (#12870)
Relates to [this
ticket](https://linear.app/vercel/issue/ECO-1144/investigate-mux-prompt-issue-in-cli).

We originally wrote the CLI integration install wizard to enable the
creation of storage products, and weren't yet certain if future product
groups would use the same structure and systems, so to be safe our CLI
redirects non-storage products to the dashboard to finish setup. The
video product from Mux uses the same underlying structures as storage,
just a different protocol.

In the future, if different integration categories continue to all share
the same underlying structure, we should be able to remove the guardrail
check around if the integration is of a product type supported by the
CLI.